+++ core/prototypes/list.py
@@ -1,14 +1,12 @@
 import logging
 from typing import Any, Optional
-
 from viur.core import db, errors, exposed, forcePost, forceSSL, securitykey, utils
 from viur.core.cache import flushCache
-from viur.core.prototypes.module import Module
+from viur.core.base.skelmodule import SkelModule
 from viur.core.skeleton import SkeletonInstance
-from viur.core.utils import currentRequest
 
 
-class List(Module):
+class List(SkelModule):
     """
         List module prototype.
 
@@ -206,7 +204,7 @@ class List(Module):
         if not self.canEdit(skel):
             raise errors.Unauthorized()
         if (len(kwargs) == 0  # no data supplied
-            or not currentRequest.get().isPostRequest  # failure if not using POST-method
+            or not utils.currentRequest.get().isPostRequest  # failure if not using POST-method
             or not skel.fromClient(kwargs)  # failure on reading into the bones
             or ("bounce" in kwargs and kwargs["bounce"] == "1")  # review before changing
         ):
@@ -242,7 +240,7 @@ class List(Module):
             raise errors.Unauthorized()
         skel = self.addSkel()
         if (len(kwargs) == 0  # no data supplied
-            or not currentRequest.get().isPostRequest  # failure if not using POST-method
+            or not utils.currentRequest.get().isPostRequest  # failure if not using POST-method
             or not skel.fromClient(kwargs)  # failure on reading into the bones
             or ("bounce" in kwargs and kwargs["bounce"] == "1")  # review before adding
         ):
@@ -310,7 +308,7 @@ class List(Module):
                     raise errors.Forbidden()
                 seoUrl = utils.seoUrlToEntry(self.moduleName, skel)
                 # Check whether this is the current seo-key, otherwise redirect to it
-                if currentRequest.get().request.path != seoUrl:
+                if utils.currentRequest.get().request.path != seoUrl:
                     raise errors.Redirect(seoUrl, status=301)
                 self.onView(skel)
                 return self.render.view(skel)
