+++ core/prototypes/tree.py
@@ -1,13 +1,12 @@
 import logging
-from typing import Any, Dict, List, Literal, Optional, Type, Union
+from typing import Any, Dict, List, Literal, Optional, Type
 from viur.core import utils, errors, conf, securitykey, db
 from viur.core import forcePost, forceSSL, exposed, internalExposed
 from viur.core.bones import KeyBone, SortIndexBone
 from viur.core.cache import flushCache
-from viur.core.prototypes.module import Module
+from viur.core.base.skelmodule import SkelModule
 from viur.core.skeleton import Skeleton, SkeletonInstance
 from viur.core.tasks import CallDeferred
-from viur.core.utils import currentRequest
 
 SkelType = Literal["node", "leaf"]
 
@@ -36,7 +35,7 @@ class TreeSkel(Skeleton):
                 db.Key.from_legacy_urlsafe(skelValues.dbEntity["parentdir"]))
 
 
-class Tree(Module):
+class Tree(SkelModule):
     """
     Tree module prototype.
 
@@ -361,7 +360,7 @@ class Tree(Module):
 
         if (len(kwargs) == 0  # no data supplied
             or not skel.fromClient(kwargs)  # failure on reading into the bones
-            or not currentRequest.get().isPostRequest
+            or not utils.currentRequest.get().isPostRequest
             or ("bounce" in kwargs and kwargs["bounce"] == "1")  # review before adding
         ):
             return self.render.add(skel)
@@ -406,7 +405,7 @@ class Tree(Module):
             raise errors.Unauthorized()
         if (len(kwargs) == 0  # no data supplied
             or not skel.fromClient(kwargs)  # failure on reading into the bones
-            or not currentRequest.get().isPostRequest
+            or not utils.currentRequest.get().isPostRequest
             or ("bounce" in kwargs and kwargs["bounce"] == "1")  # review before adding
         ):
             return self.render.edit(skel)
